Add files via upload
This is a 12 stage program that will replace failed stages automatically. After a stage is called lack of a running signal will cause the next stage in line to take over. Stages can be set up to run in any order in the queue system. Automatic rotation can replace the first stage with the last after a set runtime. Up to 4 failed stages can automatically be replace before user intervention. This was designed for heating stages but it can be used for anything it would fit.